+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 6.4.0
 
 - Queue time capture for Rack ([#2838](https://github.com/getsentry/sentry-ruby/pull/2838))